requirements.txt
sqlalchemy
pymysql
sendgrid
azure.cosmos